// interfaces/Book.ts
export interface IBook {
  title: string;
  author: string;
  publishYear: number;

}
export interface FormData {
  title: string;
  author: string;
  publishYear: number;
}
export interface UpdateDeleteResponse {
	response: string;
	status: number;
}
//   {
// 	"title":"The 5AM CLUB",
//     "author":"Robin Sharma",
//     "publishYear":"2018"
// }